New way of dealing with duplicated jobs
We now allow storing them, but at runtime they are aborted. This is
achieved by computing a checksum of the parameters hash. In fact, this
allows to detect cases where different stacks lead to the same final hash !

There are two additional job states: PARAM_CHECK, which is used when eHive
is checking whether the job has already been selected, and REDUNDAT to
track discarded jobs.